rearranged code and fixed loop detection logic
Arguemnts:
Alphabetized the parameter list to make it easier to use. Tweaked parameter
names so that there are no longer duplicates. Switches attempt to be as
similar to scamper.trace as possible.

Static Helper Methods:
Moved static helper methods into the TraceTask class to increase portability
and also remove the need for a TraceResult object to be passed into some of
the methods.

Loop Detection Algorithm:
Fixed a couple bugs that were introduced in the port from scamper.
Updated the -L parameter (loopaction) so that it counts the number of
adjacent loop *instances* that are allowed in a trace. For example,
-L of 0 will allow no adjacent interfaces (AA will halt). -L 1 will allow
one, so (AAXBB) will halt. -L will allow two, and so on.
Also removed the ability for -l to be greater than 1. This means that loop
detection has two major modes: disabled (no loop checking) and enabled.
If in enabled mode, the first time a non-adjacent loop is detected, probing
will halt. For example, ABA will halt probing.