test harness and related bug fixing
modified the traceroute classes so that they are almost entirely standalone.
this will make it easier to separate them into a library later on.
added a simulate option.
added test code for a large number of potentially risky paths.
fixed an issue with multiple replies that would probably never have been seen
in real life due to duplicate filtering at the manager level.