minor output tweaks to match the readme
fixed a number of output strings where the code did not match the documented
results. For example, the F1S result in the reverse path check needed to be
P1S as we now consider it a pass (and send it to the distance check).

noted some odd logic which results in a no-op, in the code which needs to be
revisited at some point.
it shouldn't have any effect, but would be good to tidy it eventually.